predefined cpp macros can produce unexpected results (bug 3413)
GNU cpp is predefining a number of symbols, depending on the host and target
architecture. This can produce some unexpected results: for example, the
expansion of CLIENTHOST if the host name is i386.my.domain.

The attached patch creates a new -undef option to xrdb that is passed to
cpp.

Acked-by: Gaetan Nadon <memsize@videotron.ca>
Signed-off-by: Matthieu Herrb <matthieu.herrb@laas.fr>
Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>