Allow the CPP macro to contain preprocessors with command line options (v2)
The preprocessor used for xrdb may require a command line option to
produce the desired output. For the GNU cpp this could be
'traditional-cpp' which may not be valid for other preprocessors.
Therefore support the specification of preprocessors along with
required command line arguments when using the '--with-cpp' configure
option.
Example:
   ./configure --with-cpp="/usr/bin/cpp --traditional-cpp, /usr/lib/cpp".

v2: Followed a suggestion by Julien Cristau <jcristau@debian.org>
    to allocate memory for the dup string dynamically instead of
    using a static buffer.

Signed-off-by: Egbert Eich <eich@freedesktop.org>